refactor: SectionProps 제네릭 타입 변경
