+++ fe/src/components/SearchBar/Personnel/Personnel.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import SearchSection from 'components/SearchBar/SearchSection/SearchSection';
-import { SectionProps } from 'components/SearchBar/types';
+import SearchSection from 'components/common/Section';
+import { PersonnelType, SectionProps } from 'components/SearchBar/types';
 
-export default function Personnel({ search, addSearch }: SectionProps) {
-  const { title, defaultValue, value } = search;
-  const { adult, teenager } = value;
+export default function Personnel({ info, search }: SectionProps<PersonnelType>) {
+  const { adult, teenager, child } = search;
   const guest = (adult || 0) + (teenager || 0);
+  const value = `게스트: ${guest}명${child ? `,유아:${child}명` : ''}`;
 
-  return <SearchSection title={title as string} value={guest || defaultValue} />;
+  return <SearchSection title={info.title} value={guest ? value : info.defaultValue} />;
 }
