+++ fe/src/components/SearchBar/Price/Price.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import SearchSection from 'components/SearchBar/SearchSection/SearchSection';
-import { SectionProps } from 'components/SearchBar/types';
+import SearchSection from 'components/common/Section';
+import { PriceType, SectionProps } from 'components/SearchBar/types';
 
-export default function Price({ search, addSearch }: SectionProps) {
-  const { title, defaultValue, value } = search;
-  const { minPrice, maxPrice } = value;
-  const price = minPrice || maxPrice ? `${minPrice}~${maxPrice}` : defaultValue;
+export default function Price({ info, search }: SectionProps<PriceType>) {
+  const { minPrice, maxPrice } = search;
+  const price = minPrice || maxPrice ? `${minPrice}~${maxPrice}` : info.defaultValue;
 
-  return <SearchSection title={title as string} value={price} />;
+  return <SearchSection title={info.title} value={price} />;
 }
